remove old Dagger/Anvil opt-in
